Separate out CamphoricFetcher into separate class
